feat: Add CI workflow with GitHub Actions (#1)
* feat: Add CI workflow with GitHub Actions

This commit introduces a new Continuous Integration (CI) pipeline using GitHub Actions.

The workflow is triggered on pull requests and pushes to the `main` branch. It performs the following checks to ensure code quality and stability:
- Checks for consistent code formatting.
- Runs static analysis to catch potential errors.
- Builds the web application to verify that the project compiles.

A status badge has also been added to the README.md to display the current build status.

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

* fix:github actions

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

* implement MVP app

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

---------

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>